Basically syntactic consistency

+++ 1.c
@@ -1,13 +1,16 @@
-# include <stdio.h>
+#include <stdio.h>
 
-void main() {
-  int i=1;
-  int r=0;
-  while(i<1000){
-    if(i%3==0 || i%5==0){
+void main()
+{
+  long i = 1;
+  long r = 0;
+
+  while(i<1000) {
+    if(i%3==0 || i%5==0) {
       r += i;
     }
     i++;
   }
-  printf("%d\n",r);
+
+  printf("%ld\n", r);
 }
