+++ 4.c
@@ -1,16 +1,19 @@
-# include <stdio.h>
-# include "palindromeQ.h"
+#include <stdio.h>
+#include "palindromeQ.h"
 
-void main() {
-  int i = 100;
+void main()
+{
+  long i = 100;
   long r = 0;
-  while(i<1000){
-    int j=i;
-    while(j<1000){
-      if(i*j>r && palindromeQ(i*j))r = i*j;
+
+  while(i<1000) {
+    long j = i;
+    while(j<1000) {
+      if(i*j>r && palindromeQ(i*j)) r = i*j;
       j++;
     }
     i++;
   }
-  printf("%d\n",r);
+
+  printf("%ld\n", r);
 }
